splitted the handlers and storage into separate packages
