+++ internal/handlers/handlers.go
@@ -1,0 +1,40 @@
+package handlers
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"net/http"
+	"strings"
+
+	"github.com/netzen86/collectmetrics/internal/repositories/memstorage"
+)
+
+func GetMetrics(storage *memstorage.MemStorage) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		ctx := context.Background()
+		if r.Method == http.MethodPost {
+			uri := strings.Split(r.RequestURI, "/")
+			// io.WriteString(w, r.RequestURI)
+			if len(uri) == 5 {
+				err := storage.UpdateParam(ctx, uri[2], uri[3], uri[4])
+				if uri[1] != "update" {
+					err = errors.New("wrong method use update")
+				}
+				if err != nil {
+					http.Error(w, err.Error(), http.StatusBadRequest)
+					w.WriteHeader(http.StatusBadRequest)
+				}
+				w.WriteHeader(http.StatusOK)
+				fmt.Println("!!!!", storage)
+			} else {
+				http.Error(w, "Metrics name not found!", http.StatusNotFound)
+				w.WriteHeader(http.StatusNotFound)
+			}
+			return
+		} else {
+			http.Error(w, "Use POST method!", http.StatusBadRequest)
+			w.WriteHeader(http.StatusBadRequest)
+		}
+	}
+}
