+++ internal/repositories/repositories.go
@@ -1,0 +1,7 @@
+package repositories
+
+import "context"
+
+type Repo interface {
+	UpdateParam(ctx context.Context, metricType, metricName, metricValue string) error
+}
